+++ CHANGELOG.rst
@@ -13,7 +13,7 @@ Release Notes
 - ``align_wcs`` now will raise a custom exception of type ``NotEnoughCatalogs``
   when there are not enough input catalogs to perform alignment. [#203]
  
-- ``XYXYMatch`` now will raise a custom exception of type ``MultiMatchError``
+- ``XYXYMatch`` now will raise a custom exception of type ``MatchSourceConfusionError``
   when multipe reference sources match a single input source. [#204]